+++ src/pages/product-detail/components/ProductInfo/Buttons/index.js
@@ -1,5 +1,5 @@
-import ButtonsForBuyer from "./forBuyer";
-import ButtonsForSeller from "./forSeller";
+import ButtonsForBuyer from "./Buyer";
+import ButtonsForSeller from "./Seller";
 
 const ButtonContainer = ({ isSeller, bookmark, chat, status }) => {
 	if (isSeller) return <ButtonsForSeller chat={chat} status={status} />;
